restored JSONBuilder polymorphism and fixed the bug instead of replacing it with if casts
